requirements.txt
fastapi
uvicorn[standard]
prisma
pydantic
python-dotenv
reportlab
python-barcode
passlib[bcrypt]
pyjwt
authlib
httpx
